Split endpoints based tests into different project
