+++ src/ServiceComposer.AspNetCore.Endpoints.Tests/DelegateAuthenticationSchemeOptions.cs
@@ -1,14 +1,12 @@
-﻿#if NETCOREAPP3_1
-using System;
+﻿using System;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Http;
 
-namespace ServiceComposer.AspNetCore.Tests
+namespace ServiceComposer.AspNetCore.Endpoints.Tests
 {
     class DelegateAuthenticationSchemeOptions : AuthenticationSchemeOptions
     {
         public Func<HttpRequest, Task<AuthenticateResult>> OnAuthenticate { get; set; }
     }
 }
-#endif
