+++ src/ServiceComposer.AspNetCore.Endpoints.Tests/Get_with_2_handlers.cs
@@ -1,43 +1,37 @@
-﻿#if NETCOREAPP3_1
-
-using System.Threading.Tasks;
+﻿using System.Threading.Tasks;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Routing;
 using Microsoft.Extensions.DependencyInjection;
 using Newtonsoft.Json.Linq;
 using ServiceComposer.AspNetCore.Testing;
 using Xunit;
 
-namespace ServiceComposer.AspNetCore.Tests.When_using_endpoints
+namespace ServiceComposer.AspNetCore.Endpoints.Tests
 {
-    public class Get_with_1_handler_and_1_subscriber
+    public class Get_with_2_handlers
     {
-        class TestEvent{}
-
-        class TestGetHandlerThatAppendAStringAndRaisesTestEvent : ICompositionRequestsHandler
+        class TestGetIntegerHandler : ICompositionRequestsHandler
         {
             [HttpGet("/sample/{id}")]
-            public async Task Handle(HttpRequest request)
+            public Task Handle(HttpRequest request)
             {
+                var routeData = request.HttpContext.GetRouteData();
                 var vm = request.GetComposedResponseModel();
-                vm.AString = "sample";
-
-                await vm.RaiseEvent(new TestEvent());
+                vm.ANumber = int.Parse(routeData.Values["id"].ToString());
+                return Task.CompletedTask;
             }
         }
 
-        class TestGetSubscriberThatAppendAnotherStringWhenTestEventIsRaised : ICompositionEventsSubscriber
+        class TestGetStrinHandler : ICompositionRequestsHandler
         {
             [HttpGet("/sample/{id}")]
-            public void Subscribe(ICompositionEventsPublisher publisher)
+            public Task Handle(HttpRequest request)
             {
-                publisher.Subscribe<TestEvent>((@event, request) =>
-                {
-                    var vm = request.GetComposedResponseModel();
-                    vm.AnotherString = "sample";
-                    return Task.CompletedTask;
-                });
+                var vm = request.GetComposedResponseModel();
+                vm.AString = "sample";
+                return Task.CompletedTask;
             }
         }
 
@@ -45,15 +39,15 @@ namespace ServiceComposer.AspNetCore.Tests.When_using_endpoints
         public async Task Returns_expected_response()
         {
             // Arrange
-            var client = new SelfContainedWebApplicationFactoryWithWebHost<Get_with_1_handler_and_1_subscriber>
+            var client = new SelfContainedWebApplicationFactoryWithWebHost<Get_with_2_handlers>
             (
                 configureServices: services =>
                 {
                     services.AddViewModelComposition(options =>
                     {
                         options.AssemblyScanner.Disable();
-                        options.RegisterCompositionHandler<TestGetHandlerThatAppendAStringAndRaisesTestEvent>();
-                        options.RegisterCompositionHandler<TestGetSubscriberThatAppendAnotherStringWhenTestEventIsRaised>();
+                        options.RegisterCompositionHandler<TestGetStrinHandler>();
+                        options.RegisterCompositionHandler<TestGetIntegerHandler>();
                     });
                     services.AddRouting();
                 },
@@ -75,9 +69,7 @@ namespace ServiceComposer.AspNetCore.Tests.When_using_endpoints
             var responseObj = JObject.Parse(responseString);
 
             Assert.Equal("sample", responseObj?.SelectToken("AString")?.Value<string>());
-            Assert.Equal("sample", responseObj?.SelectToken("AnotherString")?.Value<string>());
+            Assert.Equal(1, responseObj?.SelectToken("ANumber")?.Value<int>());
         }
     }
 }
-
-#endif
