--- src/ServiceComposer.AspNetCore.Tests/Properties/launchSettings.json
@@ -1,10 +1,0 @@
-{
-  "ServiceComposer.AspNetCore.Tests": {
-    "commandName": "Project",
-    "launchBrowser": false,
-    "environmentVariables": {
-      "ASPNETCORE_ENVIRONMENT": "Development"
-    },
-    "applicationUrl": "http://localhost:52569/"
-  }
-}
